add tlc5947 signal device (for https://www.adafruit.com/product/1429) add lights to each signal add signal lights tests

+++ libwinston/DigitalCentralStation.cpp
@@ -30,6 +30,6 @@ namespace winston
 
 	void DigitalCentralStation::turnoutUpdate(Turnout::Shared turnout, const Turnout::Direction direction)
 	{
-		this->signalBox->order(Command::make([turnout, direction](const unsigned long& created) -> const State { return turnout->finalizeChangeTo(direction);  }));
+		this->signalBox->order(Command::make([turnout, direction](const unsigned long long& created) -> const State { return turnout->finalizeChangeTo(direction);  }));
 	}
 };
