+++ winston-tests/winston-tests-mini.cpp
@@ -107,7 +107,7 @@ namespace winstontests
 
             auto direction = winston::Turnout::Direction::A_C;
             auto cb = std::make_shared<winston::Callback>([]() {});
-            signalBox->order(winston::Command::make([t1, direction](const unsigned long& created) -> const winston::State { return t1->finalizeChangeTo(direction);  }));
+            signalBox->order(winston::Command::make([t1, direction](const unsigned long long& created) -> const winston::State { return t1->finalizeChangeTo(direction);  }));
             //EventTurnoutFinalizeToggle::make(cb, t1, direction));
             for (int i = 0; i < 10; ++i)
                 signalBox->work();
@@ -132,7 +132,7 @@ namespace winstontests
 
             auto direction = winston::Turnout::Direction::A_B;
             auto cb = std::make_shared<winston::Callback>([]() {});
-            signalBox->order(winston::Command::make([t1](const unsigned long& created) -> const winston::State { return t1->startToggle(); }));
+            signalBox->order(winston::Command::make([t1](const unsigned long long& created) -> const winston::State { return t1->startToggle(); }));
             //signalBox->notify(winston::EventTurnoutStartToggle::make(cb, t1));
             for (int i = 0; i < 10; ++i)
                 signalBox->work();
@@ -143,7 +143,7 @@ namespace winstontests
             Assert::IsFalse(t1->traverse(winston::Track::Connection::C, onto, false));
 
             auto cb2 = std::make_shared<winston::Callback>([]() {});
-            signalBox->order(winston::Command::make([t1, direction](const unsigned long& created) -> const winston::State { return t1->finalizeChangeTo(direction); }));
+            signalBox->order(winston::Command::make([t1, direction](const unsigned long long& created) -> const winston::State { return t1->finalizeChangeTo(direction); }));
             //signalBox->notify(winston::EventTurnoutFinalizeToggle::make(cb2, t1, direction));
             for (int i = 0; i < 10; ++i)
                 signalBox->work();
@@ -176,13 +176,13 @@ namespace winstontests
             Assert::IsTrue(sBA->shows(winston::Signal::Aspect::Go));
             Assert::IsTrue(sCA->shows(winston::Signal::Aspect::Halt));
 
-            signalBox->order(winston::Command::make([t1](const unsigned long& created) -> const winston::State { return t1->finalizeChangeTo(winston::Turnout::Direction::A_C); }));
+            signalBox->order(winston::Command::make([t1](const unsigned long long& created) -> const winston::State { return t1->finalizeChangeTo(winston::Turnout::Direction::A_C); }));
             for (int i = 0; i < 10; ++i)
                 signalBox->work();
             Assert::IsTrue(sBA->shows(winston::Signal::Aspect::Halt));
             Assert::IsTrue(sCA->shows(winston::Signal::Aspect::Go));
             
-            signalBox->order(winston::Command::make([t1](const unsigned long& created) -> const winston::State { return t1->finalizeChangeTo(winston::Turnout::Direction::A_B); }));
+            signalBox->order(winston::Command::make([t1](const unsigned long long& created) -> const winston::State { return t1->finalizeChangeTo(winston::Turnout::Direction::A_B); }));
             for (int i = 0; i < 10; ++i)
                 signalBox->work();
             Assert::IsTrue(sBA->shows(winston::Signal::Aspect::Go));
